refactor: move session files into _session subpackage
