+++ src/deephaven_mcp/sessions/_session/__init__.py
@@ -1,0 +1,1 @@
+# This file marks the _session directory as a Python package.
